perf(elisp-bundle.el): Implement scanning of top-level lists in buffer
Implement the scanning and parsing of top-level s-expressions in a given
buffer. Introduce local variables for tracking top-level s-expressions
and facilitate return of buffer objects. Implement a function to find
and return the symbol provided by the given s-expressions. Modify the
looking-at for interactive expressions. Implement an alternative to
elisp-bundle--provided-feature and change the code to use the new
implementation. Additionally, modify and improve the function
elisp-bundle-include-undefined.